AI_Driven_Crop_Yield_Prediction_and_Resource_Optimization_System/requirements.txt
streamlit
pandas
matplotlib.pyplot
mysql.connector
google.generativeai
graphviz
bcrypt
mysql-connector-python
